requirements.txt
motor==3.6.0
python-dotenv==1.0.1
pymongo==4.9.2
httpx==0.27.2
ratelimit==2.2.1